requirements.txt
aerich==0.9.2
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.12.0
asyncclick==8.3.0.7
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dictdiffer==0.9.0
Flask==3.1.2
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
ollama==0.6.1
priority==2.0.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pypika-tortoise==0.6.3
pytz==2025.2
Quart==0.20.0
quart-auth==0.11.0
requests==2.32.5
tomlkit==0.13.3
tortoise-orm==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
uvicorn==0.38.0
Werkzeug==3.1.4
wsproto==1.3.2
yarl==1.22.0
